Centered Bee's Growth and Shrink and Prevented Growth that results in getting stuck

+++ Pos.py
@@ -1,7 +1,7 @@
 from Rect import Rect
 
 class Pos:
-    def __init__(self,x:int,y:int):
+    def __init__(self,x:float,y:float):
         self.x = x
         self.y = y
